Josh's work on Taylor's schedule.php page
- I got tired of having different files so I used FileZilla to steal
Taylors ;-)
- I edited the selectClasses.php page to look a little better for
testing but its nowhere near the final thing
- I heavily edited the schedule.php page to read the first value from
the array of results, read the first string in that value, and run a
query in our database to print the CRN of all classes whose subject
match the read string from the value. All this was for testing and can
be edited/expanded on request.