Edited what already existed
I changed selectClasses.php to send two arrays instead of one, and I
updated schedule.php use the values in both arrays to run a query asking
for info from the Classes table. If you type in CS 180, it gives you
detail about CS 180 and its sections in the Class table. I'll throw in
info from the other tables in a bit...